cleaning files, updating tests
