+++ wzdx/sample_files/enhanced/navjoy_wzdx copy.json
@@ -1,0 +1,71 @@
+{
+    "road_event_feed_info": {
+        "feed_info_id": "104d7746-688c-44ed-b195-2ee948bf9dfa",
+        "update_date": "2021-12-01T06:13:23Z",
+        "publisher": "CDOT",
+        "contact_name": "Ashley Nylen",
+        "contact_email": "ashley.nylen@state.co.us",
+        "version": "3.1",
+        "license": "https://creativecommons.org/publicdomain/zero/1.0/",
+        "data_sources": [
+            {
+                "data_source_id": "f5ea0ea9-46ec-4cbb-9e84-1d4084d95285",
+                "feed_info_id": "104d7746-688c-44ed-b195-2ee948bf9dfa",
+                "organization_name": "CDOT",
+                "contact_name": "Ashley Nylen",
+                "contact_email": "ashley.nylen@state.co.us",
+                "update_date": "2021-12-01T06:13:23Z",
+                "location_method": "channel-device-method",
+                "lrs_type": "lrs_type"
+            }
+        ]
+    },
+    "type": "FeatureCollection",
+    "features": [
+        {
+            "type": "Feature",
+            "properties": {
+                "road_event_id": "83b6dc60-15b5-420a-8fe7-4c7e7e3170d8",
+                "event_type": "work-zone",
+                "data_source_id": "f5ea0ea9-46ec-4cbb-9e84-1d4084d95285",
+                "start_date": "2020-12-11T16:18:33Z",
+                "end_date": "",
+                "start_date_accuracy": "estimated",
+                "end_date_accuracy": "estimated",
+                "beginning_accuracy": "estimated",
+                "ending_accuracy": "estimated",
+                "road_names": [
+                    "I-75 S"
+                ],
+                "direction": "southbound",
+                "vehicle_impact": "all-lanes-open",
+                "lanes": [],
+                "beginning_cross_street": "",
+                "ending_cross_street": "",
+                "event_status": "active",
+                "types_of_work": [],
+                "restrictions": [],
+                "description": "19-1210: Roadwork between MP 28.9 and MP 33.3",
+                "creation_date": "2021-12-01T06:13:23Z",
+                "update_date": "2021-12-01T06:06:01Z"
+            },
+            "geometry": {
+                "type": "LineString",
+                "coordinates": [
+                    [
+                        -84.099021,
+                        37.049914
+                    ],
+                    [
+                        -84.098418,
+                        37.035513
+                    ],
+                    [
+                        -84.104104,
+                        37.006754
+                    ]
+                ]
+            }
+        }
+    ]
+}
